+++ src/CustomItemRegistry/Plugin/CustomItemRegistryPlugin.cs
@@ -6,11 +6,13 @@ namespace ValheimCustomItemRegistry
 {
     [BepInPlugin(PluginGuid, PluginName, PluginVersion)]
     [BepInDependency(Jotunn.Main.ModGuid)]
+    [BepInDependency("com.ValheimModding.YamlDotNetDetector", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("com.ValheimModding.NewtonsoftJsonDetector", BepInDependency.DependencyFlags.SoftDependency)]
     public sealed class CustomItemRegistryPlugin : BaseUnityPlugin
     {
         public const string PluginGuid = "com.valheimcustomitemregistry.api";
         public const string PluginName = "Custom Item Registry";
-        public const string PluginVersion = "0.5.0";
+        public const string PluginVersion = "0.6.0";
 
         internal static ManualLogSource Log { get; private set; }
 
@@ -23,6 +25,8 @@ namespace ValheimCustomItemRegistry
 
             harmony = new Harmony(PluginGuid);
             harmony.PatchAll();
+
+            CustomItemRegistry.LoadItemPacks();
 
             Logger.LogInfo($"{PluginName} {PluginVersion} loaded");
         }